+++ hydromt/data_catalog.py
@@ -468,6 +468,7 @@ class DataCatalog(object):
         source_names: List = [],
         unit_conversion: bool = True,
         meta: Dict = {},
+        append: bool = False,
     ) -> None:
         """Export a data slice of each dataset and a data_catalog.yml file to disk.
 
@@ -481,11 +482,16 @@ class DataCatalog(object):
             Start and end date of period of interest. By default the entire time period
             of the dataset is returned.
         source_names: list, optional
-            List of source names to export
+            List of source names to export, by default None in which case all sources are exported.
+            Specific variables can be selected by appending them to the source name in square brackets.
+            For example, to export all variables of 'source_name1' and only 'var1' and 'var2' of 'source_name'
+            use source_names=['source_name1', 'source_name2[var1,var2]']
         unit_conversion: boolean, optional
             If False skip unit conversion when parsing data from file, by default True.
         meta: dict, optional
             key-value pairs to add to the data catalog meta section, such as 'version', by default empty.
+        append: bool, optional
+            If True, append to existing data catalog, by default False.
         """
         data_root = abspath(data_root)
         if not os.path.isdir(data_root):
@@ -505,10 +511,16 @@ class DataCatalog(object):
         else:
             sources = copy.deepcopy(self.sources)
 
+        # read existing data catalog if it exists
+        fn = join(data_root, "data_catalog.yml")
+        if isfile(fn) and append:
+            self.logger.info(f"Appending existing data catalog {fn}")
+            sources_out = DataCatalog(fn).sources
+        else:
+            sources_out = {}
+
         # export data and update sources
-        sources_out = {}
         for key, source in sources.items():
-            postfix = f"_{'_'.join(source_vars[key])}" if key in source_vars else ""
             try:
                 # read slice of source and write to file
                 self.logger.debug(f"Exporting {key}.")
@@ -519,7 +531,7 @@ class DataCatalog(object):
                     source.unit_add = {}
                 fn_out, driver = source.to_file(
                     data_root=data_root,
-                    data_name=f"{key}{postfix}",
+                    data_name=key,
                     variables=source_vars.get(key, None),
                     bbox=bbox,
                     time_tuple=time_tuple,
@@ -540,14 +552,17 @@ class DataCatalog(object):
                 source.filesystem = "local"
                 source.kwargs = {}
                 source.rename = {}
-                sources_out[f"{key}{postfix}"] = source
+                if key in sources_out:
+                    self.logger.warning(
+                        f"{key} already exists in data catalog and is overwritten."
+                    )
+                sources_out[key] = source
             except FileNotFoundError:
                 self.logger.warning(f"{key} file not found at {source.path}")
 
         # write data catalog to yml
         data_catalog_out = DataCatalog()
         data_catalog_out._sources = sources_out
-        fn = join(data_root, "data_catalog.yml")
         data_catalog_out.to_yml(fn, root="auto", meta=meta)
 
     def get_rasterdataset(